pyproject.toml
[tool.poetry]
name = "nonebot-plugin-repeater"
version = "3.2.0"
description = "A plugin based on NoneBot2, auto +1 in group."
authors = ["Kl1nge5", "UKMeng <ukmengse@gmail.com>"]
license = "GPLv3"
readme = "README.md"
packages = [{include = "nonebot_plugin_repeater"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
nonebot-plugin-alconna = "^0.25.0"
nonebot-plugin-session = "^0.1.0"
# 想要用red最好用dev分支
# nonebot-plugin-session = {git = "https://github.com/noneplugin/nonebot-plugin-session.git", rev = "dev"}
nonebot-plugin-send-anything-anywhere = "^0.3.0"

[tool.poetry.group.dev.dependencies]
nonebot2 = {extras = ["fastapi", "httpx", "webscokets"], version = "^2.1.0"}
nonebot-adapter-red = "^0.5.0"
nonebot-adapter-onebot = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
adapters = [
    { name = "RedProtocol", module_name = "nonebot.adapters.red" },
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = [ "nonebot_plugin_repeater"]
plugin_dirs = []
builtin_plugins = ["echo"]